+++ NEWS.md
@@ -8,6 +8,7 @@ version increments reflect backwards-incompatible API changes, not necessarily b
 * Set `options('huxtable.latex_use_fontspec')` to `TRUE` to use portable font 
   names in TeX documents, with the LaTeX "fontspec" package.
 * Bugfix: attributes were being copied wrongly in subset assignment of huxtables.
+* Bugfix: text colors in `hux_logo()`.
 
 # huxtable 4.5.0